+++ mullvad-daemon/src/lib.rs
@@ -505,6 +505,7 @@ where
 
         let account_history = account_history::AccountHistory::new(
             &cache_dir,
+            &settings_dir,
             rpc_handle.clone(),
             core_handle.remote.clone(),
         )